feat(logger): add client-side logger and utilities
