--- lib/logger/instances/index.ts
@@ -1,11 +1,0 @@
-export {
-  logger,
-  apiLogger,
-  stravaLogger,
-  authLogger,
-  dbLogger,
-  createBrowserLogger,
-  createLogger,
-  createRequestLogger,
-  logApiRequest,
-} from './instances';
